requirements.txt
Flask==2.3.3
requests==2.31.0
pandas==2.2.3
pinecone-client==3.2.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
streamlit==1.29.0
tqdm==4.66.1
sentence-transformers==2.7.0
groq
pytesseract